Remove task-specific code from common library
