--- src/Kononov/Common/Mesh.cpp
@@ -1,1 +1,0 @@
-#include "Mesh.hpp"
