--- src/Kononov/Common/Shader.cpp
@@ -1,2 +1,0 @@
-#include "Shader.hpp"
-
